Move networking responsiblities to separate class.
